Functionality was moved to UserLayerHandler
